# chdrawappbackend/src/main/java/org/example/drawapp/dto/Credentials.java
package org.example.drawapp.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record Credentials(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 8) String password
) {
}